Fix to put the form on a different panel so it can be hidden when no results are shown. Trello 174
