+++ patientview-parent/radar/src/main/java/org/patientview/radar/web/pages/patient/AddPatientPage.html
@@ -93,6 +93,8 @@
 
 
      <div wicket:id="selectPatientPanel"></div>
+     <div wicket:id="createPatientPanel"></div>
+
