+++ patientview-parent/radar/src/main/java/org/patientview/radar/web/panels/CreatePatientPanel.html
@@ -1,0 +1,8 @@
+<wicket:panel>
+
+    <form wicket:id="patientCreationForm">
+        <input class="saveBtn" type="button" wicket:id="create" value="Create New Patient"/>
+    </form>
+
+
+</wicket:panel>
