+++ patientview-parent/radar/src/main/java/org/patientview/radar/web/panels/SelectPatientPanel.html
@@ -68,9 +68,4 @@
 
     </form>
 
-    <form wicket:id="patientCreationForm">
-        <input class="saveBtn" type="button" wicket:id="create" value="Create New Patient"/>
-    </form>
-
-
 </wicket:panel>
